refactor: remove system agents and double independent agents
- Removed all system agents (reduced from 25 to 0)
- Doubled independent agents (increased from 25 to 50)
- Maintains same total initial agent population of 50
- Simplifies simulation by focusing on independent agent behavior
