+++ agents/__init__.py
@@ -1,0 +1,2 @@
+from .independent_agent import IndependentAgent
+from .system_agent import SystemAgent
