+++ config.yaml
@@ -3,8 +3,8 @@ width: 100
 height: 100
 
 # Agent settings
-system_agents: 25
-independent_agents: 25
+system_agents: 0
+independent_agents: 50
 initial_resource_level: 12
 max_population: 300
 starvation_threshold: 0
